# docs/swagger.yaml
basePath: /
definitions:
  helper.BaseResponse:
    properties:
      data: {}
      errors: {}
      status:
        type: integer
    type: object
  helper.ExampleErrorResponse:
    properties:
      email:
        example: cannot be blank
        type: string
      message:
        example: user bad request
        type: string
      password:
        example: cannot be blank
        type: string
    type: object
...
  modelphoto.Request:
    properties:
      caption:
        example: Beautiful Photo
        type: string
      photo_url:
        example: https://example.com/photo.jpg
        type: string
      title:
        example: A Photo
        type: string
    type: object
  modelphoto.Response:
    properties:
      caption:
        example: Beautiful Photo
        type: string
      created_at:
        example: "2021-11-03T01:52:41.035Z"
        type: string
      id:
        example: 1
        type: integer
      photo_url:
        example: https://example.com/photo.jpg
        type: string
      title:
        example: A Photo
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  modelphoto.ResponseGet:
    properties:
      caption:
        example: Beautiful Photo
        type: string
      created_at:
        example: "2021-11-03T01:52:41.035Z"
        type: string
      id:
        example: 1
        type: integer
      photo_url:
        example: https://example.com/photo.jpg
        type: string
      title:
        example: A Photo
        type: string
      updated_at:
        example: "2021-11-03T01:52:41.035Z"
        type: string
      user:
        properties:
          email:
            example: test@example.com
            type: string
          username:
            example: jhondoe
            type: string
        type: object
    type: object
  modelphoto.ResponseUpdate:
    properties:
      caption:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  modelsocialmedia.Request:
    properties:
      name:
        type: string
      social_media_url:
        type: string
    type: object
  modelsocialmedia.Response:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      social_media_url:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  modelsocialmedia.ResponseList:
    properties:
      created_at:
        example: "2021-11-03T01:52:41.035Z"
        type: string
      id:
        example: 1
        type: integer
      name:
        example: jhondoe
        type: string
      social_media_url:
        example: https://example.com/url
        type: string
      updated_at:
        example: "2021-11-03T01:52:41.035Z"
        type: string
      user:
        properties:
          id:
            example: 1
            type: integer
          profile_image_url:
            example: https://example.com/photo.jpg
            type: string
          username:
            example: jhondoe
            type: string
        type: object
      user_id:
        example: 1
        type: integer
    type: object
  modelsocialmedia.ResponseListWrapper:
    properties:
      social_medias:
        items:
          $ref: '#/definitions/modelsocialmedia.ResponseList'
        type: array
    type: object
...
info:
  contact: {}
  description: This is API for completing final project 2 hacktiv8
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Hacktiv8 Mygram API
  version: "1.0"
paths:
...
  /photos:
    get:
      consumes:
      - application/json
      description: GetPhotos a photo
      parameters:
      - description: Bearer + user token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: SUCCESS
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/modelphoto.ResponseGet'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
        "401":
          description: Unauthorization
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
      summary: GetPhotos a photo
      tags:
      - photos
    post:
      consumes:
      - application/json
      description: Create new photo
      parameters:
      - description: Bearer + user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/modelphoto.Request'
      produces:
      - application/json
      responses:
        "201":
          description: CREATED
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/modelphoto.Response'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
        "401":
          description: Unauthorization
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
      summary: Create new photo
      tags:
      - photos
  /photos/:photoID:
    delete:
      consumes:
      - application/json
      description: Delete a photo
      parameters:
      - description: Bearer + user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of the photo
        in: path
        name: photoID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: SUCCESS
          schema:
            $ref: '#/definitions/helper.BaseResponse'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
        "401":
          description: Unauthorization
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
      summary: Delete a photo
      tags:
      - photos
    put:
      consumes:
      - application/json
      description: Update a photo
      parameters:
      - description: Bearer + user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of the photo
        in: path
        name: photoID
        required: true
        type: integer
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/modelphoto.Request'
      produces:
      - application/json
      responses:
        "200":
          description: SUCCESS
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/modelphoto.ResponseUpdate'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
        "401":
          description: Unauthorization
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
      summary: Update a photo
      tags:
      - photos
  /socialmedias:
    get:
      consumes:
      - application/json
      description: Get all social media
      parameters:
      - description: Bearer + user token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: SUCCESS
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/modelsocialmedia.ResponseListWrapper'
              type: object
      summary: Get all social media
      tags:
      - socialmedias
    post:
      consumes:
      - application/json
      description: Create social media
      parameters:
      - description: Bearer + user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/modelsocialmedia.Request'
      produces:
      - application/json
      responses:
        "201":
          description: CREATED
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/modelsocialmedia.Response'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
      summary: Create new social media
      tags:
      - socialmedias
  /socialmedias/:socialmediaid:
    delete:
      consumes:
      - application/json
      description: Delete by id social media
      parameters:
      - description: Bearer + user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of the social media
        in: path
        name: socialmediaid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: SUCCESS
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/modelsocialmedia.Response'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
        "404":
          description: Record not found
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
      summary: Delete by id social media
      tags:
      - socialmedias
    put:
      consumes:
      - application/json
      description: Update by id social media
      parameters:
      - description: Bearer + user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of the social media
        in: path
        name: socialmediaid
        required: true
        type: integer
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/modelsocialmedia.Request'
      produces:
      - application/json
      responses:
        "200":
          description: SUCCESS
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/modelsocialmedia.Response'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
        "404":
          description: Record not found
          schema:
            allOf:
            - $ref: '#/definitions/helper.BaseResponse'
            - properties:
                errors:
                  $ref: '#/definitions/helper.ExampleErrorResponse'
              type: object
      summary: Update by id social media
      tags:
      - socialmedias
...
swagger: "2.0"